+++ sumologic/provider.go
@@ -111,6 +111,8 @@ func Provider() terraform.ResourceProvider {
 			"sumologic_content_permission":                       resourceSumologicPermissions(),
 			"sumologic_local_file_source":                        resourceSumologicLocalFileSource(),
 			"sumologic_log_search":                               resourceSumologicLogSearch(),
+			"sumologic_metrics_search":                           resourceSumologicMetricsSearch(),
+			"sumologic_rum_source":                               resourceSumologicRumSource(),
 		},
 		DataSourcesMap: map[string]*schema.Resource{
 			"sumologic_cse_log_mapping_vendor_product": dataSourceCSELogMappingVendorAndProduct(),