+++ sumologic/resource_sumologic_slo.go
@@ -164,8 +164,9 @@ func resourceSumologicSLO() *schema.Resource {
 		Schema: map[string]*schema.Schema{
 
 			"name": {
-				Type:     schema.TypeString,
-				Required: true,
+				Type:         schema.TypeString,
+				Required:     true,
+				ValidateFunc: validation.StringDoesNotContainAny("/"),
 			},
 			"description": {
 				Type:     schema.TypeString,